pyproject.toml
[project]
name = "agentic-investor"
dynamic = ["version"]
description = "A Model Context Protocol server for an investor agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anysqlite>=0.0.5",
    "hishel>=1.0.0",
    "html5lib>=1.1",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "mcp[cli]>=1.19.0",
    "pandas>=2.3.3",
    "pytrends>=4.9.2",
    "requests-cache>=1.2.1",
    "tenacity>=9.1.2",
    "yfinance[nospam]>=0.2.66",
]
authors = [
    { name = "jfox", email = "howsitgoing@email.com" }
]
keywords = ["investment", "finance", "trading", "agent", "mcp", "model context protocol", "broker", "robinhood", "wallstreetbets"]
classifiers = [
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Office/Business :: Financial",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]

[project.optional-dependencies]
ta = [
    "ta-lib>=0.6.8",
]
alpaca = [
    "alpaca-py>=0.43.1",
]

[dependency-groups]
dev = [
    "pydantic-ai>=1.9.0",
    "python-dotenv>=1.2.1",
]

[project.scripts]
agentic-investor = "agentic_investor.server:mcp.run"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agentic_investor"]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "agentic_investor/_version.py"
